+++ packages/terra-worklist-data-grid/tests/jest/ColumnResizeHandle.test.jsx
@@ -1,5 +1,5 @@
-/* eslint-disable import/no-extraneous-dependencies */
 import React from 'react';
+/* eslint-disable-next-line import/no-extraneous-dependencies */
 import { mountWithIntl } from 'terra-enzyme-intl';
 import ColumnResizeHandle from '../../src/subcomponents/ColumnResizeHandle';